Deleted the UserByUserType class. Moved the method in UserManager where it should belong. + some small changes.
