+++ src/BackEnd/Restaurant/Order.java
@@ -15,7 +15,7 @@ import static FrontEnd.Validators.formatDecimalNumber;
 public class Order {
     private List<OrderedDish> orderedDishes;
     private double totalPrice = 0;
-    private boolean isPaid;
+    private final boolean isPaid;
     private OrderStatus orderStatus;
     private final int tableNumber;
     private long orderNumber;  // Very long in the DB. Int might not be long enough. Will be created from the DB.
